requirements.txt
tensorflow-macos
tensorflow-metal
opencv-python
numpy
pandas
matplotlib
jupyter
scipy
seaborn
scikit-learn